Add tricubic interpolation, fix thread safety, and improve robustness Tricubic interpolation: - Add triCubicInterp and triCubicGradient to ScalarGrid with fallback   to trilinear near boundaries - Add cubicInterpGradient utility (analytic derivative of cubic kernel) - Expose tricubic interpolation on VectorGrid and LevelSet - Add InterpolationOrder enum to FieldAdvector for linear vs cubic - Level set advection and findSurface now use cubic interpolation - LevelSet::normal() gains option to use cubic gradient Thread-safety fixes: - Replace bare bool flags shared across TBB parallel_for with   std::atomic<bool> in GeometricMultigridOperators,   GeometricMultigridPoissonSolver, and GeometricPressureProjection Supersampling volume computation fix: - Replace floating-point loop bounds with integer loop counters in   computeSupersampleVolumes and computeSupersampledFaceVolumes to   avoid accumulation drift - Clamp supersampled volumes to [0, 1] Timer modernization: - Replace platform-specific timing (QueryPerformanceCounter /   gettimeofday) with std::chrono::high_resolution_clock Minor fixes and cleanup: - Fix SparseUniformGrid using std::pow(TILE_SIZE, 3) instead of   integer multiplication - Fix ScalarGrid default constructor not initializing mySampleType,   myBorderType, myCellOffset - Add missing default assert cases in ScalarGrid and VectorGrid   switch statements - Fix NotchedDiskField::operator() parameter from float to double - Add const to EulerianLiquidSimulator::maxVelocityMagnitude() - Fix shadowed variable name in multigrid symmetry test - Remove unused variables (volumeSamples, worldPoint, colours array) - Fix double semicolon in UniformGrid::reset - Fix namespace closing comment and missing newline in VectorGrid.h New unit tests: - Cubic interpolation and gradient convergence tests for all ScalarGrid   sample types - Convergence tests for lerp, trilerp, trilerpGradient, cubicInterp,   and cubicInterpGradient utilities - Update testJitterMesh to use triCubicInterp to match findSurface

+++ Library/SimTools/TestVelocityFields.h
@@ -173,7 +173,7 @@ class NotchedDiskField
 public:
 
     // Procedural velocity field
-    Vec3d operator()(float, const Vec3d& pos) const
+    Vec3d operator()(double, const Vec3d& pos) const
     {
         return Vec3d((PI / 314.) * (50.0 - pos[1]), (PI / 314.) * (pos[0] - 50.0), 0);
     }
